Docs: Remove nobody certificate in getting started
The nobody certificate is a somewhat confusing concept and
I don't want to introduce it to early.

The use of invite-codes forced us to introduce this concept early.
However, we shifted to developer certificates now which makes onboarding
smoother.

I've slightly altered the flow of the document a little bit.

The concepts of authentication and certificates are closely related.
I've decided to bundle this information into a single section instead of
spreading it out over the document.